Clean up createHandle references
