+++ src/lang.ts
@@ -1,6 +1,7 @@
 import { PropertyEvent, Observer } from './observers/interfaces';
 import * as ObjectObserver from './observers/ObjectObserver';
 import has from './has';
+import { Handle, EventObject } from './interfaces';
 
 var slice = Array.prototype.slice;
 var hasOwnProperty = Object.prototype.hasOwnProperty;
@@ -212,4 +213,21 @@ export function partial(targetFunction: (...args: any[]) => any, ...suppliedArgs
 
 		return targetFunction.apply(this, args);
 	};
+}
+
+export function createHandle(destructor: () => void): Handle {
+	return {
+		destroy: function () {
+			this.destroy = function () {};
+			destructor.call(this);
+		}
+	};
+}
+
+export function createCompositeHandle(...handles: Handle[]): Handle {
+	return createHandle(function () {
+		for (var i = 0, handle: Handle; (handle = handles[i]); ++i) {
+			handle.destroy();
+		}
+	});
 }
